Switched all ids from long to int

+++ back-wealth-app/src/main/java/com/wealth_management_system/BackWealthApp/domain/Asset.java
@@ -13,7 +13,7 @@ public class Asset {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-	private long id;
+	private int id;
 	private String type;
 	private String descr;
 	private double origValue;
@@ -33,12 +33,12 @@ public class Asset {
 
 	
 	//Get id
-	public long id() {
+	public int id() {
 		return id;
 	}
 	
 	//Set id
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
